First stab at cleaning up ERA5 WRF Xray.
